rename typec driver
